+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 1. [](#improved)
    * `Login::register()` now validates all the provided built-in user fields, not just username
+   * Improved user field validation
+   * Do not validate optional `fullname` field, it's already handled in the registration form
 
 # v3.4.1
 ## 02/17/2021